add datatable in contractdetail

+++ Migrations/ApplicationDbContextModelSnapshot.cs
@@ -92,11 +92,11 @@ namespace ExcelFilesCompiler.Migrations
 
             modelBuilder.Entity("ExcelFilesCompiler.Models.ContractDetails", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<string>("AddedBy")
                         .HasColumnType("nvarchar(max)");
